RUM-2704: Move classes in domain to net
