+++ features/dd-sdk-android-session-replay/src/test/kotlin/com/datadog/android/sessionreplay/SessionReplayTest.kt
@@ -9,7 +9,7 @@ package com.datadog.android.sessionreplay
 import com.datadog.android.api.feature.FeatureSdkCore
 import com.datadog.android.sessionreplay.forge.ForgeConfigurator
 import com.datadog.android.sessionreplay.internal.SessionReplayFeature
-import com.datadog.android.sessionreplay.internal.domain.SegmentRequestFactory
+import com.datadog.android.sessionreplay.internal.net.SegmentRequestFactory
 import fr.xgouchet.elmyr.annotation.Forgery
 import fr.xgouchet.elmyr.annotation.StringForgery
 import fr.xgouchet.elmyr.junit5.ForgeConfiguration
